requirements.txt
torch
pandas
numpy
scikit-learn
fastapi
uvicorn
streamlit
requests
joblib
matplotlib
pillow